Read-in and averaging functions complete. Now begin restructure.
This is the last commit before starting a new plan: I plan to delegate
the data needs of NLevel to an object of a certain new class, its
provisionary name being LevelDataProvider. The subclasses of NLevel will
determine what kind of LevelDataProvider is used, and implement any
exceptions to the generic NLevel approach.