Cut off integration over line when the rest is estimated negligible.
A very conservative estimate of the rest of the integral over the
spectrum times the voigt profile is made (maximum of spectrum * last
known voigt profile * width of the remaining frequency interval).
Because the voigt profile is a descending function when moving away from
the center, and because we multiplied with them maximum of the spectrum,
we can be sure that if this value is insignificant, then the real
integrandum is even less significant.