Rework addToBinned to make more sense
No longer calculates all the bin edges before going into the final for
loop. Instead, they are calculated on the fly, and the way the for loop
iteration looks is also much more readable.

Also, the integrand is evaluated at the bin centers, instead of at the
bin edges. This also makes more sense, because that's the frequency for
which the bins represent a measurement of the radiation field.